Remove styleguide rule about line length
The line length is not relevant for the rendered version, we don't currently seem to enforce this rule, and it hinders editing.

Currently we have each paragraph on one line. This makes it possible to use the automatic word wrap of your favorite editor. An alternative would be to have newlines in the file to split a paragraph over multiple lines. This would make it easier to read the file in a text editor, but every time you edit the paragraph you would have to do line wrapping again.