Update writeup_Diving_into_mobile_cryptography_using_dynamic_instrumentation_with_frida.md
Add note regarding requirement for the local workstation's version of frida version to match the version of frida-server that is pushed to the test device that's hosting the application under test (emulator).
Clarify frida command description that connects to the test device, specifies the application under test and loads the hooking script.