Update Services subtitle
The wording makes it sound like services always run on new threads but they do not necessarily run off of the UI thread. Services are good for background tasks but not the ideal way of running async tasks. Also the  "Activities are executed in the main app thread." sentence does not fit in the paragraph and does not make sense at that part of the text.