Update external_links.js to fix link behavior on MASVS page
This PR modifies the code to enhance the user experience on the MASVS page (https://mas.owasp.org/MASVS/) by ensuring that external links open in a new tab, similar to the behavior on the MASTG page (https://mas.owasp.org/MASTG/). 

This change involves updating the checking if the path starts with '/MASTG' or '/MASVS' and applying the necessary attributes to external links accordingly.