Enhance crypto security allowing only the no-arguments constructor of SecureRandom (MASTG-TEST-0016)
To be specific in the test documentation it would be better to provide a unique clear guideline instead of a suggestion.
Instead of preferring a constructor, make it the best and only option to be compliant.

During the review, it has also been updated the MSC02-J link from the legacy domain `www.securecoding.cert.org` to the one it redirects to: `wiki.sei.cmu.edu`.

Note: the revision has been discussed and supported during the review of the Mobile App Profile security document hosted by the Linux Foundation's App Defense Alliance.